Enable nodes with weird return types.
For some EDSLs (Aplite, Opaleye, etc.), the run function needs to have a more
complicated type than DSL a -> CIO a. For instance, an Opaleye computation of
type Query (Column PGText) will return a value of type [String] when executed.
This commit builds support for this type of conversion into the framework
itself, removing the need to wrap such nodes in extra boilerplate code.